+++ tests/test_dimmer.py
@@ -116,7 +116,7 @@ class TestDimmer:
         )
 
         # Change the mode to "on"
-        assert device.switch_on()
+        device.switch_on()
         assert device.status == STATUS.ON
         assert device.is_on
 
@@ -129,7 +129,7 @@ class TestDimmer:
         )
 
         # Change the mode to "off"
-        assert device.switch_off()
+        device.switch_off()
         assert device.status == STATUS.OFF
         assert not device.is_on